Bugfixes/cleanup

+++ app/Http/Controllers/ExtraController.php
@@ -2,68 +2,51 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\ExtraRequest;
 use App\Models\Computer;
 use App\Models\Extra;
-use App\Models\User;
 use App\Models\Worksheet;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 
 class ExtraController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        //
-    }
-
     /**
      * Show the form for creating a new resource.
      */
     public function create(Request $request)
     {
-        if (Auth::check()) {
-            return view('layouts.menu', ["navUrls" => User::getNavUrls(true), "userUrls" => Auth::user()->getUserUrls(), "connected_worksheet" => Worksheet::findOrFail($request["worksheet"]), "connected_computer" => Computer::findOrFail($request["computer"])]);
-        } else {
-            return redirect(route('home'));
+        $worksheet = Worksheet::findOrFail($request["worksheet"]);
+        if($worksheet["final"] == true){
+            return redirect(route('worksheet.show', $worksheet->id));
         }
+        return view('layouts.menu', [
+            "connected_worksheet" => $worksheet,
+            "connected_computer" => Computer::findOrFail($request["computer"])
+        ]);
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(ExtraRequest $request)
     {
-        $validated = $request->validate([
-            "worksheet_id" => "required|exists:worksheets,id",
-            "computer_id" => "required|exists:computers,id",
-            "manufacturer" => "required|string",
-            "type" => "required|string",
-            "serial_number" => "required|string|unique:extras,serial_number",
-        ]);
+        $worksheet = Worksheet::findOrFail($request["worksheet_id"]);
+        if($worksheet["final"] == true){
+            return redirect(route('worksheet.show', $worksheet->id));
+        }
 
         $extra = new Extra();
-        $extra["manufacturer"] = $validated["manufacturer"];
-        $extra["type"] = $validated["type"];
-        $extra["serial_number"] = $validated["serial_number"];
+        $extra->manufacturer = $request["manufacturer"];
+        $extra->type = $request["type"];
+        $extra->serial_number = $request["serial_number"];
 
         $extra->save();
 
-        $extra->computer()->attach($validated['computer_id'], [
-            'worksheet_id' => $validated['worksheet_id'],
+        $extra->computer()->attach($request['computer_id'], [
+            'worksheet_id' => $request['worksheet_id'],
         ]);
 
-        return redirect(route('computer.show', $validated["computer_id"]));
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-        //
+        return redirect(route('computer.show', $request["computer_id"]));
     }
 
     /**
@@ -71,33 +54,34 @@ class ExtraController extends Controller
      */
     public function edit(Request $request, string $id)
     {
-        if (Auth::check()) {
-            return view('layouts.menu', ["navUrls" => User::getNavUrls(true), "userUrls" => Auth::user()->getUserUrls(), "extra" =>  Extra::findOrFail($id), "connected_worksheet" => Worksheet::findOrFail($request["worksheet"]), "connected_computer" => Computer::findOrFail($request["computer"])]);
-        } else {
-            return redirect(route('home'));
+        $worksheet = Worksheet::findOrFail($request["worksheet"]);
+        if($worksheet["final"] == true){
+            return redirect(route('worksheet.show', $worksheet->id));
         }
+
+        return view('layouts.menu', [
+            "extra" =>  Extra::findOrFail($id),
+            "connected_worksheet" => $worksheet,
+            "connected_computer" => Computer::findOrFail($request["computer"])
+        ]);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(ExtraRequest $request, Extra $extra)
     {
-        $validated = $request->validate([
-            "worksheet_id" => "required|exists:worksheets,id",
-            "computer_id" => "required|exists:computers,id",
-            "manufacturer" => "required|string",
-            "type" => "required|string",
-        ]);
+        $worksheet = $extra->worksheet()->first();
+        if($worksheet->final == true){
+            return redirect(route('worksheet.show', $worksheet->id));
+        }
 
-        $extra = Extra::findOrFail($id);
-
-        $extra["manufacturer"] = $validated["manufacturer"];
-        $extra["type"] = $validated["type"];
+        $extra->manufacturer = $request["manufacturer"];
+        $extra->type = $request["type"];
 
         $extra->save();
 
-        return redirect(route('computer.show', $validated["computer_id"]));
+        return redirect(route('computer.show', $request["computer_id"]));
     }
 
     /**
@@ -106,6 +90,10 @@ class ExtraController extends Controller
     public function destroy(string $id)
     {
         $extra = Extra::findOrFail($id);
+        $worksheet = $extra->worksheet;
+        if($worksheet["final"] == true){
+            return redirect(route('worksheet.show', $worksheet->id));
+        }
 
         $computer_id = $extra->computer[0]->id;
         $extra->computer()->detach();
